What sparked these edits was thinking through whether or not Project SAFE is a DApp?
There seems to be a bit of a contradiction between point 1 of the definition of a DApp and the project-safe Whitepaper:

The application must be completely open-source, it must operate autonomously, with no entity controlling the majority of its tokens, and its data and records of operation must be cryptographically stored in a public, decentralized block chain.

The two contradictions are bolded below - the SAFE network doesn't use a blockchain and the transaction records can be very private.

Additionally, as a fully decentralised network, the SAFE approach allows transactions to be made and confirmed at network speed (under a second in some cases). This is due to a distributed Transaction Manager as opposed to a blockchain. In the SAFE network, a transaction management system can be linked or not. Bitcoin uses a linked blockchain (the chain term) that allows traversal of all transactions from the network start. SAFE has chosen an unlinked approach to the blockchain. Each user's account information is held by the group of nodes closest to it (according to the XOR address distance). The Transaction Manager only holds a temporary receipt object during the transaction procedure among users. This temporary receipt can be stored permanently allowing proof of the transaction to be maintained or destroyed immediately after the transaction is completed, leaving no trace on the network. In addition to allowing instant transfers of coins, this mechanism also allows an escrow model (a third party acts as moderator to resolve the payment dispute). This escrow mechanism is a core component of the currency.

One problem I foresee is that Project SAFE won't be the only one to strive to provide the users of their application with increased privacy nor will they be the only one to continue to develop new forms of decentralized databases. Per my reading of the current DApp definition this would rule out all of these projects from being considered as DApps.

What if the definition was instead focused on the governing bodies (projects/foundations/etc) operating in a public manner and leveraging a bitcoin-like decentralized database.

Disclaimer! This is in fact, in line with my long held personal world view that there should be total transparency at the core (governments or in our case projects/foundations) lots of transparency in the middle (companies or the in our case the applications) and opaqueness at the fringes (individuals or in our case the use of those applications by individuals).

Here a rough draft of how we could update it:

The application must be completely open-source, it must operate autonomously, with no entity controlling the majority of its tokens, and the operations, actions, & decisions of any governing bodies (such as a foundation to distribute tokens for development) must be cryptographically stored in a public, decentralized database such as a block chain.

A couple cosmetic points of feedback:

1 - I think we want to change the order of the DApp definition.  In point 1 tokens are referenced but haven't been defined it yet. I think if you swapped 1 & 2 it would clear things up quite a bit.

2 - I also think leading with tokens is strong b/c tokenization is the big idea here and the subsequent points are how to issue tokens and then manage the project in a decentralized manner using them.

3 - Lastly, I think we may want to use DApp throughout the definition instead of application.  We've got a bit of a language hurdle to overcome - want to start beating 'DApp' into people's brains whenever possible.

Happy to discuss further! -Dustin